Eventmanager - stabilised more for error-cases
Build: 40629

Changes:
  - Eventmanager: EventManagerScript - functions who got passed to the eventmanager are now run in protected mode.
They print out an error-message in case of an error and the accompanying event will be paused from that moment on.
This should make that more stable overall, as the other eventcheckings are not stopped due one faulty function of one eventcheck.

Replacing the function and restarting the eventchecking should be possible, so debugging it on runtime should work.

In other news: EventChecking now works as often as it was requested(used to be done in every defer-cycle of the eventmanager, which wasted computing-resources)