attribute-additions
previous/next chapter-attributes can now have a path name, that can be followed.
So you can set the next chapter to be part of a certain path, that can be followed through.

For instance: a podcast about Apollo 11 can have a path of chapters under the pathname "technical stuff", so it's possible to just listen to an order of chapters about technical stuff and leaving out all the other chapters.

Or another path about called "astronauts", which only let's the listener hear all chapters about the astronauts only, leaving out the rest.