Chapter/Shownote get Guids now
Chapters and Shownotes get unique guids now, that can be references later on.
This is more stable, when a shownote/chapter moves its position(due mispositioning during first export). The newly exported file has the shownote/chapter in a different position, but the guid stays the same.
So if before reupload of the episode, some other podcast already referenced the shownote/chapter, the reference stays intact, which would not be the case when using the position as unique identifier.